// libs/features/contributor/src/lib/shared/enums/contributor-action.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum ContributorActionEnum {
  TERMINATE = 'CONTRIBUTOR.TERMINATE',
  TRANSFER = 'CONTRIBUTOR.TRANSFER',
  CANCEL = 'CONTRIBUTOR.CANCEL',
  MODIFY = 'CONTRIBUTOR.MODIFY',
  ADD_CONTRACT = 'CONTRIBUTOR.ADD-CONTRACT',
  CONTRACT_DETAILS = 'CONTRIBUTOR.CONTRACT-DETAILS',
  RE_ACTIVATE = 'CONTRIBUTOR.RE-ACTIVATE',
  MODIFY_JOINING_DATE = 'CONTRIBUTOR.MODIFY-JOINING-DATE',
  MODIFY_LEAVING_DATE = 'CONTRIBUTOR.MODIFY-LEAVING-DATE',
  CANCEL_ENGAGEMENT = 'CONTRIBUTOR.CANCEL-CON.CANCEL-ENGAGEMENT',
  VIEW_CONTRACT = 'CONTRIBUTOR.VIEW-CONTRACTS',
  VIEW_WAGE_BREAKUP = 'CONTRIBUTOR.VIEW-WAGE-BREAKUP',
  TERMINATE_ENGAGEMENT = 'CONTRIBUTOR.TERMINATE-ENGAGEMENT',
  VIEW_VIC_BILL_DASH = 'CONTRIBUTOR.VIEW-BILL-DASHBOARD-SCREEN',
  MODIFY_COVERAGE = 'CONTRIBUTOR.MODIFY-COVERAGE',
  RAISE_OBJECTION = 'CONTRIBUTOR.RAISE-OBJECTION',
  ML_CANCEL_REQUEST = 'CONTRIBUTOR.MEDICAL-INSURANCE.CANCEL-REQUEST',
  ML_MODIFY_ENDURANCE_RATION = 'CONTRIBUTOR.MEDICAL-INSURANCE.MODIFY-ENDURANCE-RATION'
}